backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-jose
pyyaml
redis
pyfilters
pyjwt
sqlalchemy
snowflake-sqlalchemy
pydantic
phonenumbers
snowflake-id
passlib[bcrypt]
pydantic[email]